requirements.txt
streamlit==1.28.1
google-generativeai==0.3.2
PyMuPDF==1.23.8
python-dotenv==1.0.0
langchain==0.0.350
langchain-google-genai==0.0.5
tiktoken==0.5.1
reportlab==4.0.4 